Move class implementation from header to cpp file on AudioCapture.h and Listener.h. Fix warnings.
